+++ app/sitemap-products.xml/route.ts
@@ -7,7 +7,7 @@ export async function GET() {
 
   const siteUrl = process.env.SITE_URL;
 
-  sdk.configureForNext(locale);
+  sdk.defaultConfigure(locale);
 
   const fields = [] as SiteMapField[];