+++ app/sitemap-static.xml/route.ts
@@ -6,7 +6,7 @@ export async function GET() {
 
   const siteUrl = process.env.SITE_URL;
 
-  sdk.configureForNext(locale);
+  sdk.defaultConfigure(locale);
 
   const fields = [] as SiteMapField[];